[ADD&FIX] 23/10/15 (23:04)
이미지 폴더이름 member->user로 수정
프로필이미지 등록 수정
마이페이지-메인 프로필이미지 보이게 수정
이메일인증 관련 파일 추가
이메일 인증 구현 완료

+++ src/main/java/com/bringbring/user/domain/User.java
@@ -27,6 +27,6 @@ public class User {
 	private String userProfileName;
 	private String userProfileRename;
 	private String userProfilePath;
-	private int userProfileLength;
+	private long userProfileLength;
 	private char isUserDeleted;
 }
